backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.23
pymysql==1.1.0
python-dotenv==1.0.1
google-cloud-bigquery==3.13.0
google-auth==2.26.2
groq==0.9.0
cryptography==41.0.7
httpx>=0.26.0
